requirements.txt
rich>=12.0.0
requests>=2.25.1
gitpython>=3.1.30
pyyaml>=6.0
typer>=0.7.0
pathspec>=0.9.0
beautifulsoup4>=4.11.1
slack-sdk>=3.19.5
python-dotenv>=0.19.0
tqdm>=4.66
gitignore_parser>=0.1.12 